smart-study-buddy/requirements.txt
streamlit
endee==0.1.10
pymupdf
requests
python-dotenv